+++ node/cli/src/service.rs
@@ -370,7 +370,7 @@ pub fn new_full_base(
             offchain_db.local_storage_set(
                 sp_core::offchain::StorageKind::PERSISTENT,
                 b"rotate_keyshares",
-                &format!("{}/validator/rotate_keyshares", endpoint).into_bytes(),
+                &format!("{}/rotate_network_key", endpoint).into_bytes(),
             );
             offchain_db.local_storage_set(
                 sp_core::offchain::StorageKind::PERSISTENT,