+++ pallets/attestation/src/mock.rs
@@ -71,8 +71,6 @@ impl pallet_attestation::Config for Test {
     type RuntimeEvent = RuntimeEvent;
     type WeightInfo = ();
     type Randomness = TestPastRandomness;
-    type KeyProvider = Staking;
-    type AttestationQueue = Staking;
 }
 
 parameter_types! {
@@ -312,14 +310,13 @@ impl Randomness<H256, BlockNumber> for TestPastRandomness {
 }
 parameter_types! {
   pub const MaxEndpointLength: u32 = 3;
-  pub const MaxPendingAttestations: u32 = 4;
 }
 impl pallet_staking_extension::Config for Test {
+    type AttestationHandler = ();
     type Currency = Balances;
     type MaxEndpointLength = MaxEndpointLength;
     type Randomness = TestPastRandomness;
     type RuntimeEvent = RuntimeEvent;
-    type MaxPendingAttestations = MaxPendingAttestations;
     type WeightInfo = ();
     type PckCertChainVerifier = pallet_staking_extension::pck::MockPckCertChainVerifyer;
 }